clean code修改
# message auto-generated for no-merge-commit merge:
merge clean_code into master

clean code修改

Created-by: msprofDeveloper

Match-id-7df8472a2b8f00e09fb5b2b4f9de09aa0723c4ef

+++ collector/dvvp/job_wrapper/src/prof_job.cpp
@@ -1599,7 +1599,7 @@ int ProfAicpuJob::Init(const SHARED_PTR_ALIA<CollectionJobCfg> cfg)
     if (!DrvChannelsMgr::instance()->ChannelIsValid(collectionJobCfg_->comParams->devId, channelId_)) {
         MSPROF_LOGW("Channel is invalid, devId:%d, channelId:%d", collectionJobCfg_->comParams->devId,
             static_cast<int>(channelId_));
-        eventAttr_.deviceId = collectionJobCfg_->comParams->devId;
+        eventAttr_.deviceId = static_cast<uint32_t>(collectionJobCfg_->comParams->devId);
         eventAttr_.grpName = eventGrpName_;
         int32_t ret = profDrvEvent_.SubscribeEventThreadInit(&eventAttr_);
         return ret;
